Remove legacy pool.dataset.permission endpoint
This functionality has been superseded by filesystem ACL and
permissions endpoints since TrueNAS 12. Endpoint is no longer
used by the webui and so it is safe to remove.
